requirements.txt
torch>=2.0
torchvision>=0.15
streamlit>=1.30
pillow
numpy
pydicom
scikit-image
opencv-python
tqdm
pandas
streamlit>=1.30
plotly>=5.15
matplotlib>=3.7
seaborn>=0.12